+++ NEWS.md
@@ -33,6 +33,7 @@
 - RStudio no longer includes spurious chunk markers from Copilot completion results (#13686)
 - Added new commands which toggle the maximized state of Console, Source, and Tab Group quadrants [Accessibility] (#7022)
 - RStudio now prints a warning if an overly-long line of console input is truncated when sending to R (#14420)
+- Updated zlib to 1.3.1 (rstudio/rstudio-pro#5971)
 
 #### Posit Workbench
 - Show custom project names on Workbench homepage (rstudio-pro#5589)